HAAR-3818: delete the authorisation consent when client is deleted

+++ src/test/kotlin/uk/gov/justice/digital/hmpps/authorizationapi/integration/ClientsInterfaceControllerIntTest.kt
@@ -25,6 +25,7 @@ import uk.gov.justice.digital.hmpps.authorizationapi.adapter.AuthService
 import uk.gov.justice.digital.hmpps.authorizationapi.adapter.ServiceDetails
 import uk.gov.justice.digital.hmpps.authorizationapi.data.model.AuthorizationConsent
 import uk.gov.justice.digital.hmpps.authorizationapi.data.model.AuthorizationConsent.AuthorizationConsentId
+import uk.gov.justice.digital.hmpps.authorizationapi.data.model.Client
 import uk.gov.justice.digital.hmpps.authorizationapi.data.model.ClientType
 import uk.gov.justice.digital.hmpps.authorizationapi.data.model.Hosting
 import uk.gov.justice.digital.hmpps.authorizationapi.data.model.MfaAccess
@@ -195,7 +196,7 @@ class ClientsInterfaceControllerIntTest : IntegrationTestBase() {
     fun `delete success single client version only`() {
       whenever(oAuthClientSecretGenerator.generate()).thenReturn("external-client-secret")
       whenever(oAuthClientSecretGenerator.encode("external-client-secret")).thenReturn("encoded-client-secret")
-      givenANewClientExistsWithClientId("test-test")
+      val client = givenANewClientExistsWithClientId("test-test")
 
       webTestClient.delete().uri("/base-clients/test-test/clients/test-test")
         .headers(setAuthorisation(roles = listOf("ROLE_OAUTH_ADMIN")))
@@ -205,6 +206,7 @@ class ClientsInterfaceControllerIntTest : IntegrationTestBase() {
       assertNull(clientRepository.findClientByClientId("test-test"))
       assertNull(clientDeploymentRepository.findClientDeploymentByBaseClientId("test-test"))
       assertFalse(clientConfigRepository.findById("test-test").isPresent)
+      verifyAuthorizationConsentRecordNotPresent(client.id, client.clientId)
 
       verify(telemetryClient).trackEvent(
         "AuthorizationApiClientDeleted",
@@ -219,12 +221,14 @@ class ClientsInterfaceControllerIntTest : IntegrationTestBase() {
       whenever(oAuthClientSecretGenerator.encode("external-client-secret")).thenReturn("encoded-client-secret")
       whenever(oAuthClientSecretGenerator.encode("external-client-secret-2")).thenReturn("encoded-client-secret-2")
 
-      givenANewClientExistsWithClientId("test-test")
+      val baseClient = givenANewClientExistsWithClientId("test-test")
 
       webTestClient.post().uri("/base-clients/test-test/clients")
         .headers(setAuthorisation(roles = listOf("ROLE_OAUTH_ADMIN")))
         .exchange()
         .expectStatus().isOk
+
+      val duplicateClient = clientRepository.findClientByClientId("test-test-1")
 
       webTestClient.delete().uri("/base-clients/test-test/clients/test-test-1")
         .headers(setAuthorisation(roles = listOf("ROLE_OAUTH_ADMIN")))
@@ -232,14 +236,15 @@ class ClientsInterfaceControllerIntTest : IntegrationTestBase() {
         .expectStatus().isOk
 
       assertNull(clientRepository.findClientByClientId("test-test-1"))
+      verifyAuthorizationConsentRecordNotPresent(duplicateClient!!.id, "test-test-1")
 
-      val baseClient = clientRepository.findClientByClientId("test-test")
-      val clientDeployment = clientDeploymentRepository.findClientDeploymentByBaseClientId("test-test")
+      val clientDeployment = clientDeploymentRepository.findClientDeploymentByBaseClientId("test-test")!!
       val clientConfig = clientConfigRepository.findById("test-test")
 
       assertNotNull(baseClient)
       assertNotNull(clientDeployment)
       assertTrue(clientConfig.isPresent)
+      verifyAuthorities(baseClient.id, baseClient.clientId, "ROLE_CURIOUS_API", "ROLE_VIEW_PRISONER_DATA", "ROLE_COMMUNITY")
 
       clientDeploymentRepository.delete(clientDeployment)
       clientConfigRepository.delete(clientConfig.get())
@@ -252,7 +257,7 @@ class ClientsInterfaceControllerIntTest : IntegrationTestBase() {
       )
     }
 
-    private fun givenANewClientExistsWithClientId(clientId: String) {
+    private fun givenANewClientExistsWithClientId(clientId: String): Client {
       webTestClient.post().uri("/base-clients")
         .headers(setAuthorisation(roles = listOf("ROLE_OAUTH_ADMIN")))
         .body(
@@ -295,9 +300,11 @@ class ClientsInterfaceControllerIntTest : IntegrationTestBase() {
         .exchange()
         .expectStatus().isOk
 
-      assertNotNull(clientRepository.findClientByClientId("test-test"))
-      assertNotNull(clientDeploymentRepository.findClientDeploymentByBaseClientId("test-test"))
-      assertTrue(clientConfigRepository.findById("test-test").isPresent)
+      val client = clientRepository.findClientByClientId(clientId)
+      assertNotNull(client)
+      assertNotNull(clientDeploymentRepository.findClientDeploymentByBaseClientId(clientId))
+      assertTrue(clientConfigRepository.findById(clientId).isPresent)
+      return client!!
     }
   }
 
@@ -424,8 +431,8 @@ class ClientsInterfaceControllerIntTest : IntegrationTestBase() {
         .exchange()
         .expectStatus().isOk
 
-      val duplicatedClient = clientRepository.findClientByClientId("test-client-id-1")
-      val duplicatedClient2 = clientRepository.findClientByClientId("test-client-id-2")
+      val duplicatedClient = clientRepository.findClientByClientId("test-client-id-1")!!
+      val duplicatedClient2 = clientRepository.findClientByClientId("test-client-id-2")!!
 
       webTestClient.post().uri("/base-clients/test-client-id/clients")
         .headers(setAuthorisation(roles = listOf("ROLE_OAUTH_ADMIN")))
@@ -462,10 +469,10 @@ class ClientsInterfaceControllerIntTest : IntegrationTestBase() {
       assertThat(registeredClientAdditionalInformation.getJiraNumber(duplicatedClient.clientSettings)).isEqualTo(registeredClientAdditionalInformation.getJiraNumber(originalClient.clientSettings))
       assertThat(registeredClientAdditionalInformation.getDatabaseUserName(duplicatedClient.clientSettings)).isEqualTo(registeredClientAdditionalInformation.getDatabaseUserName(originalClient.clientSettings))
 
-      val authorizationConsent = authorizationConsentRepository.findByIdOrNull(AuthorizationConsentId(originalClient.id, originalClient.clientId))
-      val duplicateCateAuthorizationConsent = authorizationConsentRepository.findByIdOrNull(AuthorizationConsentId(duplicatedClient.id, duplicatedClient.clientId))
+      val authorizationConsent = authorizationConsentRepository.findByIdOrNull(AuthorizationConsentId(originalClient.id, originalClient.clientId))!!
+      val duplicateCateAuthorizationConsent = authorizationConsentRepository.findByIdOrNull(AuthorizationConsentId(duplicatedClient.id, duplicatedClient.clientId))!!
 
-      assertThat(duplicateCateAuthorizationConsent?.authorities).isEqualTo(authorizationConsent?.authorities)
+      assertThat(duplicateCateAuthorizationConsent.authorities).isEqualTo(authorizationConsent.authorities)
 
       authorizationConsentRepository.delete(duplicateCateAuthorizationConsent)
 
@@ -1744,7 +1751,7 @@ class ClientsInterfaceControllerIntTest : IntegrationTestBase() {
   }
 
   private fun verifyAuthorizationConsentRecordNotPresent(id: String, clientId: String) {
-    val authorizationConsent = authorizationConsentRepository.findById(AuthorizationConsent.AuthorizationConsentId(id, clientId))
+    val authorizationConsent = authorizationConsentRepository.findById(AuthorizationConsentId(id, clientId))
     assertFalse(authorizationConsent.isPresent)
   }
 }
